// lambda/activateVehicle/models/activation.model.ts
export interface ActivationDocument {
    activation_id: string;
    champion_id: string;
    champion_uuid_id: string;
    contractPage: string;
    contractStatus: string;
    documentStatus: string;
    drivingLicense: string;
    financierInfo: string;
    healthInsurance: string;
    helmetNumber: number;
    lastUpdateTime: string;
    nameOfFleetOfficer: string;
    parent_activation_id: null;
    platformInfo: string;
    prospect_id: string;
    serviceType: string;
    vehicleOptions: string;
    vehicle_id: string;
    prospectLocation: string;
    vehicleLocation: string;
    contractPic: string;
}